requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
redis
pymysql
flask-cors
flask-bcrypt
apscheduler